Added scene transitions
Added GlobalScenes to the autoload
Added scene transitions to existing levels
changed a bit of the way that the scenetransition script determines which scene to load.
added a saved scene to the save file.
Adjusted the way that the main menu determines to load a new scene or a saved scene